Extend the range to 64 bits.

+++ demo.hpp
@@ -1,1 +1,1 @@
-unsigned fibonacci(unsigned index);
+unsigned long fibonacci(unsigned index);
